refactor: :recycle: NavMenuOption type 중복 삭제
