+++ app/src/pages/About.tsx
@@ -4,7 +4,7 @@ export const AboutPage = () => {
   return (
     <>
       <Helmet>
-        <title>프로젝트 소개 | 42Stat</title>
+        <title>서비스 소개 | 42Stat</title>
       </Helmet>
       AboutPage
     </>
